requirements.txt
torch~=2.9.0+cu126
transformers~=4.57.1
pandas~=2.3.3
torchvision~=0.24.0+cu126
pillow~=11.3.0
tqdm~=4.67.1